add_roxygen_mask {styler}	R Documentation

Add the roxygen mask to code

Description

This function compares text with initial_text to make sure a mask is only added to roxygen comments, not ordinary comments

Usage


add_roxygen_mask(text, initial_text, example_type)

Arguments

text
Character vector with code.

initial_text
The roxygen code example to style with mask and potentially ordinary comments.

example_type
Either 'examples' or 'examplesIf'.

[Package styler version 1.10.3 Index]