alignment_ensure_trailing_comma {styler}	R Documentation

Ensure last pd has a trailing comma

Description

Must be after alignment_ensure_no_closing_brace() because if it comes after alignment_ensure_trailing_comma(), the last expression would not be a brace, which would make removal complicated.

Usage


alignment_ensure_trailing_comma(pd_by_line)

Arguments

pd_by_line
A list, each element corresponding to a potentially incomplete parse table that represents all token from one line.

[Package styler version 1.10.3 Index]