alignment_has_correct_spacing_around_comma {styler}	R Documentation

Check if spacing around comma is correct

Description

At least one space after comma, none before, for all but the last comma on the line

Usage


alignment_has_correct_spacing_around_comma(pd_sub)

Arguments

pd_sub
The subset of a parse table corresponding to one line.

[Package styler version 1.10.3 Index]