assert_transformers {styler}	R Documentation

Assert the transformers

Description

Actually only assert name and version of style guide in order to make sure caching works correctly.

Usage


assert_transformers(transformers)

Arguments

transformers
A list of transformer functions that operate on flat parse tables.

[Package styler version 1.10.3 Index]