bind_with_child {styler}	R Documentation

Bind a parse table with one of its children

Description

Bind a parse table with one of its children and return parse table, ordered according to the appearance of the tokens.

Usage


bind_with_child(pd_nested, pos)

Arguments

pd_nested
A nested parse table.

pos
The position of the child to bind.

[Package styler version 1.10.3 Index]