cache_activate {styler}	R Documentation

Activate or deactivate the styler cache

Description

Helper functions to control the behavior of caching. Simple wrappers around base::options().

Usage


cache_activate(cache_name = NULL, verbose = !getOption("styler.quiet", FALSE))

cache_deactivate(verbose = !getOption("styler.quiet", FALSE))

Arguments

cache_name
The name of the styler cache to use. If NULL, the option "styler.cache_name" is considered which defaults to the version of styler used.

verbose
Whether or not to print an informative message about what the function is doing.

See Also

Other cache managers: cache_clear(), cache_info(), caching

[Package styler version 1.10.3 Index]