cache_find_path {styler}	R Documentation

Where is the cache?

Description

Finds the path to the cache and creates it if it does not exist.

Usage


cache_find_path(cache_name = NULL)

Arguments

cache_name
The name of the styler cache to use. If NULL, the option "styler.cache_name" is considered which defaults to the version of styler used.

[Package styler version 1.10.3 Index]