cache_info {styler}	R Documentation

Show information about the styler cache

Description

Gives information about the cache. Note that the size consumed by the cache will always be displayed as zero because all the cache does is creating an empty file of size 0 bytes for every cached expression. The inode is excluded from this displayed size but negligible.

Usage


cache_info(cache_name = NULL, format = "both")

Arguments

cache_name
The name of the cache for which to show details. If NULL, the active cache is used. If none is active the cache corresponding to the installed styler version is used.

format
Either "lucid" for a summary emitted with base::cat(), "tabular" for a tabular summary from base::file.info() or "both" for both.

See Also

Other cache managers: cache_activate(), cache_clear(), caching

[Package styler version 1.10.3 Index]