cache_more_specs {styler}	R Documentation

Create more specs

Description

Syntactic sugar for creating more specs. This is useful when we want to add more arguments (because we can search for this function in the source code).

Usage


cache_more_specs(include_roxygen_examples, base_indention)

[Package styler version 1.10.3 Index]