communicate_summary {styler}	R Documentation

Communicate the summary of styling

Description

Communicate the summary of styling

Usage


communicate_summary(changed, ruler_width)

Arguments

changed
Boolean with indicating for each file whether or not it has been changed.

ruler_width
Integer used to determine the width of the ruler.

[Package styler version 1.10.3 Index]