communicate_warning {styler}	R Documentation

Communicate a warning if necessary

Description

If round trip verification was not possible, issue a warning to review the changes carefully.

Usage


communicate_warning(changed, transformers)

Arguments

changed
Boolean with indicating for each file whether or not it has been changed.

transformers
The list of transformer functions used for styling. Needed for reverse engineering the scope.

[Package styler version 1.10.3 Index]