compute_parse_data_nested {styler}	R Documentation

Obtain a nested parse table from a character vector

Description

Parses text to a flat parse table and subsequently changes its representation into a nested parse table with nest_parse_data().

Usage


compute_parse_data_nested(
  text,
  transformers = tidyverse_style(),
  more_specs = NULL
)

Arguments

text
The text to parse.

transformers
Passed to cache_make_key() to generate a key.

more_specs
Passed to cache_make_key() to generate a key.

Value

A nested parse table. See tokenize() for details on the columns of the parse table.

Examples


code <- "
ab     <- 1L # some comment
abcdef <- 2L
"
writeLines(code)
compute_parse_data_nested(code)

[Package styler version 1.10.3 Index]