construct_tree {styler}	R Documentation

Construct paths of a tree object given the paths of *-in.R files

Description

Construct paths of a tree object given the paths of *-in.R files

Usage


construct_tree(in_paths, suffix = "_tree")

Arguments

in_paths
Character vector of *-in.R files.

suffix
Suffix for the tree object.

[Package styler version 1.10.3 Index]