construct_vertical {styler}	R Documentation

Construct an object of class vertical

Description

Sole purpose of the class vertical is to have a print method that aligns the output vertically.

Usage


construct_vertical(x)

Arguments

x
A character vector or an object of class "vertical".

[Package styler version 1.10.3 Index]