create_tree {styler}	R Documentation

Create a tree from text

Description

Create a tree representation from a text.

Usage


create_tree(text, structure_only = FALSE)

Arguments

text
A character vector.

structure_only
Whether or not create a tree that represents the structure of the expression without any information on the tokens. Useful to check whether two structures are identical.

Value

A data frame.

[Package styler version 1.10.3 Index]