dir_without_. {styler}	R Documentation

dir(), but without dot-prefix and different defaults

Description

When using dir(), you can set full.names = FALSE, but then you can only pass a character vector of length one as path to not loose the information about where the files are. This function solves that case. It's needed when one wants to standardize paths to use set operations on them, i.e. when the user supplied input does not have a dot prefix. See 'Examples'.

Usage


dir_without_.(path, recursive = TRUE, ...)

Arguments

path
A path.

...
Passed to base::dir().

Details

For different defaults, see dir_without_._one.

See Also

set_and_assert_arg_paths

Examples


setdiff("./file.R", "file.R") # you want to standardize first.

[Package styler version 1.10.3 Index]