emulate_rd {styler}	R Documentation

Convert roxygen comments to Rd code

Description

We leverage roxygen2 workhorse function roxygen2::roc_proc_text() if our input contains character that have to be escaped. Since this is an expensive operation, we opt out of it and perform a simple remove_roxygen_mask() when there are no characters to escape.

Usage


emulate_rd(roxygen)

[Package styler version 1.10.3 Index]