env_current {styler}	R Documentation

The elements that are added to this environment are:

Description

The elements that are added to this environment are:

Usage


env_current

Format

An object of class environment of length 0.

Details

parser_version: Needed to dispatch between parser versions, see parser_version_set() for details.

stylerignore: A data frame with parse data containing tokens that fall within a stylerignore sequence. This is used after serializing the flattened parse table to apply the initial formatting to these tokens. See stylerignore for details.

any_stylerignore: Whether there is any stylerignore marker. The idea is to check early in the runtime if this is the case and then if so, take as many short-cuts as possible. See stylerignore for details.

[Package styler version 1.10.3 Index]