find_blank_lines_to_next_expr {styler}	R Documentation

Find blank lines

Description

What number of line breaks lay between the expressions?

Usage


find_blank_lines_to_next_expr(pd_nested)

Arguments

pd_nested
A nested parse table.

Value

The line number on which the first token occurs.

[Package styler version 1.10.3 Index]