generate_test_samples {styler}	R Documentation

Generate a comprehensive collection test cases for comment / insertion interaction Test consist of if / if-else / if-else-if-else cases, paired with various line-break and comment configurations. Used for internal testing.

Description

Generate a comprehensive collection test cases for comment / insertion interaction Test consist of if / if-else / if-else-if-else cases, paired with various line-break and comment configurations. Used for internal testing.

Usage


generate_test_samples()

Value

The function is called for its side effects, i.e. to write the test cases to *-in.R files that can be tested with test_collection(). Note that a few of the test cases are invalid and need to be removed / commented out manually.

[Package styler version 1.10.3 Index]