Get the transformer functions for flat styling

Get the transformer functions for flat styling

get_transformers_flat(strict = TRUE, start_comments_with_one_space = FALSE)

Arguments

strict
A logical value indicating whether a set of strict or not so strict transformer functions should be returned.

start_comments_with_one_space
Whether or not comments should start with only one space (see start_comments_with_space()).

Value

A list of transformer functions that operate on flat parse tables.

See also

Other obtain transformers: get_transformers_nested