hash_standardize {styler}	R Documentation

Standardize text for hashing

Description

Make sure text after styling results in the same hash as text before styling if it is indeed identical. This function expects trailing blank lines in text were removed prior to passing it to this function.

Usage


hash_standardize(text)

Arguments

text
A character vector.

[Package styler version 1.10.3 Index]