identify_start_to_stop_of_roxygen_examples_from_text {styler}	R Documentation

Figure out where code examples start and stop

Description

Finds the sequence from start to stop of the lines in text that are code examples in roxygen comments.

Usage


identify_start_to_stop_of_roxygen_examples_from_text(text)

Arguments

text
A text consisting of code and/or roxygen comments.

[Package styler version 1.10.3 Index]