is_double_indent_function_declaration {styler}	R Documentation

Is the function declaration double indented?

Description

Assumes you already checked if it's a function with is_function_declaration. It is double indented if the first token after the first line break that is a "SYMBOL_FORMALS".

Usage


is_double_indent_function_declaration(pd, indent_by = 2L)

Arguments

pd
A parse table.

indent_by
How many spaces of indention should be inserted after operators such as '('.

[Package styler version 1.10.3 Index]