is_insufficiently_parsed_string {styler}	R Documentation

Identify strings that were not fully parsed

Description

Identifies strings that were not fully parsed due to their vast length.

Usage


is_insufficiently_parsed_string(pd)

Arguments

pd
A parse table.

Details

The meaning of the variable is_problematic_string in the source code changes from "all strings" to "all problematic strings", is partly misleading and this approach was chosen for performance reasons only.

[Package styler version 1.10.3 Index]