map_filetype_to_pattern {styler}	R Documentation

Map the file type to a corresponding regular expression

Description

Map the file type to a corresponding regular expression

Usage


map_filetype_to_pattern(filetype)

Arguments

filetype
The file type to map to a regex.

Examples


styler:::map_filetype_to_pattern(c(".rMd", "R"))

[Package styler version 1.10.3 Index]