next_non_comment {styler}	R Documentation

Find the index of the next or previous non-comment in a parse table.

Description

Find the index of the next or previous non-comment in a parse table.

Usage


next_non_comment(pd, pos)

previous_non_comment(pd, pos)

Arguments

pd
A parse table.

pos
The position of the token to start the search from.

See Also

Other third-party style guide helpers: pd_is, scope_normalize()

Examples


code <- "a <- # hi \n x %>% b()"
writeLines(code)
pd <- compute_parse_data_nested(code)
child <- pd$child[[1]]
previous_non_comment(child, 4L)
next_non_comment(child, 2L)

[Package styler version 1.10.3 Index]