parse_safely {styler}	R Documentation

Save parsing from text

Description

Parses text safely, i.e. throws an informative error if EOL style does not match LF or indicates the exact position where the parsing failed. Note that we can only detect wrong EOL style if it occurs on the first line already.

Usage


parse_safely(text, ...)

Arguments

text
Text to parse.

...
Parameters passed to base::parse().

Examples


try(styler:::parse_safely("a + 3 -4 -> x\r\n glück + 1"))
# This cannot be detected as a EOL style problem because the first
# line ends as expected with \n
try(styler:::parse_safely("a + 3 -4 -> x\nx + 2\r\n glück + 1"))

styler:::parse_safely("a + 3 -4 -> \n glück + 1")

[Package styler version 1.10.3 Index]