parse_transform_serialize_r {styler}	R Documentation

Parse, transform and serialize text

Description

Wrapper function for the common three operations.

Usage


parse_transform_serialize_r(
  text,
  transformers,
  base_indention,
  warn_empty = TRUE,
  is_roxygen_code_example = FALSE
)

Arguments

text
The text to parse.

transformers
Passed to cache_make_key() to generate a key.

base_indention
Integer scalar indicating by how many spaces the whole output text should be indented. Note that this is not the same as splitting by line and add a base_indention spaces before the code in the case multi-line strings are present. See 'Examples'.

warn_empty
Whether or not a warning should be displayed when text does not contain any tokens.

is_roxygen_code_example
Is code a roxygen examples block?

See Also

parse_transform_serialize_roxygen()

[Package styler version 1.10.3 Index]