parse_tree_must_be_identical {styler}	R Documentation

Check whether a round trip verification can be carried out

Description

If scope was set to "line_breaks" or lower (compare tidyverse_style()), we can compare the expression before and after styling and return an error if it is not the same.

Usage


parse_tree_must_be_identical(transformers)

Arguments

transformers
The list of transformer functions used for styling. Needed for reverse engineering the scope.

[Package styler version 1.10.3 Index]