print.vertical {styler}	R Documentation

Print styled code

Description

Print styled code

Usage


## S3 method for class 'vertical'
print(
  x,
  ...,
  colored = getOption("styler.colored_print.vertical"),
  style = prettycode::default_style()
)

Arguments

x
A character vector, one element corresponds to one line of code.

...
Not currently used.

colored
Whether or not the output should be colored with prettycode::highlight().

style
Passed to prettycode::highlight().

[Package styler version 1.10.3 Index]