remove_roxygen_header {styler}	R Documentation

Remove roxygen header

Description

Can't simply remove the element with the regex because it may happen that the roxygen tag is on the same line as its contents start.

Usage


remove_roxygen_header(text)

Examples


#' @examples c(1, 2)

[Package styler version 1.10.3 Index]