rep_char {styler}	R Documentation

Repeat elements of a character vector times times and collapse it

Description

Repeat elements of a character vector times times and collapse it

Usage


rep_char(char, times)

Arguments

char
A character vector.

times
an integer giving the number of repetitions.

Value

A character vector.

[Package styler version 1.10.3 Index]