save_after_styling_is_active {styler}	R Documentation

Heuristic to see if a file styled with the addin should be saved or not.

Description

Using the R option "styler.save_after_styling" and if unset, checks legacy method via environment variable save_after_styling.

Usage


save_after_styling_is_active()

[Package styler version 1.10.3 Index]