separate_chunks {styler}	R Documentation

Separate chunks within Rmd and Rnw contents

Description

Identifies and separates the code and text chunks (the latter includes non-R code) within an Rmd or Rnw file, and returns these separately.

Usage


separate_chunks(lines, filetype)

Arguments

lines
A character vector of lines from an Rmd or Rnw file.

filetype
A string indicating the filetype - either 'Rmd' or 'Rnw'.

[Package styler version 1.10.3 Index]