serialize_parse_data_flattened {styler}	R Documentation

Serialize flattened parse data

Description

Collapses a flattened parse table into character vector representation.

Usage


serialize_parse_data_flattened(flattened_pd, indent_character = "")

Arguments

flattened_pd
A flattened parse table.

indent_character
The character that is used for indention. We strongly advise for using spaces as indention characters.

[Package styler version 1.10.3 Index]