set_and_assert_arg_filetype {styler}	R Documentation

Set the file type argument

Description

Sets and asserts the file type argument to a standard format for further internal processing.

Usage


set_and_assert_arg_filetype(filetype)

Arguments

filetype
A character vector with file types to convert to the internal standard format.

Examples


styler:::set_and_assert_arg_filetype("rMd")
try(styler:::set_and_assert_arg_filetype("xyz"))

[Package styler version 1.10.3 Index]