Prettify arbitrary R code

Performs various substitutions in all .R files in a directory. Carefully examine the results after running this function!

style_dir(path = ".", ..., style = tidyverse_style,
  transformers = style(...), filetype = "R", recursive = TRUE,
  exclude_files = NULL)

Arguments

path
Path to a directory with files to transform.

...
Arguments passed on to the style function.

style
A function that creates a style guide to use, by default tidyverse_style() (without the parentheses). Not used further except to construct the argument transformers. See style_guides() for details.

transformers
A set of transformer functions. This argument is most conveniently constructed via the style argument and .... See 'Examples'.

filetype
Vector of file extensions indicating which filetypes should be styled. Case is ignored, and the . is optional, e.g. c(".R", ".Rmd") or c("r", "rmd").

recursive
A logical value indicating whether or not files in subdirectories of path should be styled as well.

exclude_files
Character vector with paths to files that should be excluded from styling.

Value

Invisibly returns a data frame that indicates for each file considered for styling whether or not it was actually changed.

Warning

This function overwrites files (if styling results in a change of the code to be formatted). It is strongly suggested to only style files that are under version control or to create a backup copy.

See also

Other stylers: style_file, style_pkg, style_text, styler_addins

Examples

# NOT RUN {
style_dir(file_type = "r")
# }