style_space_around_math_token {styler}	R Documentation

Style spacing around math tokens

Description

Style spacing around math tokens

Usage


style_space_around_math_token(strict, zero, one, pd_flat)

Arguments

strict
Whether the rules should be applied strictly or not.

zero
Character vector of tokens that should be surrounded with zero spaces.

one
Character vector with tokens that should be surrounded by at least one space (depending on strict = TRUE in the styling functions style_text() and friends). See 'Examples'.

pd_flat
A nest or a flat parse table.

[Package styler version 1.10.3 Index]