style_space_around_token {styler}	R Documentation

Set spacing of token to a certain level

Description

Set the spacing of all tokens in pd_flat to level if strict = TRUE or to at least to level if strict = FALSE.

Usage


style_space_around_token(
  pd_flat,
  strict,
  tokens,
  level_before,
  level_after = level_before
)

Arguments

pd_flat
A nest or a flat parse table.

strict
Whether the rules should be applied strictly or not.

tokens
Character vector with tokens that should be styled.

level_before, level_after
Scalar indicating the amount of spaces that should be inserted around the tokens on the left and right position respectively.

[Package styler version 1.10.3 Index]