test_dry {styler}	R Documentation

Test the dry argument

Description

Test the dry argument

Usage


test_dry(path, styler, styled = FALSE)

Arguments

path
A path to pass to the styler.

styler
A function that takes path, typically a user exposed styler function that has side effects, like style_file().

[Package styler version 1.10.3 Index]