test_transformer {styler}	R Documentation

Transforming test input with a transformer function

Description

These functions can be used as inputs for test_collection() and transform_and_check().

Usage


style_empty(text, base_indention = 0L)

style_op(text, base_indention = 0L)

Arguments

text
A character vector to transform.

Details

As inputs for test_collection(), we can also use top-level functions such as style_text().

Functions

style_empty(): Nest and unnest text without applying any transformations but remove EOL spaces and indention due to the way the serialization is set up.

style_op(): Transformations for indention based on operators

[Package styler version 1.10.3 Index]