text_to_flat_pd {styler}	R Documentation

Creates a flat parse table with minimal initialization

Description

Creates a flat parse table with minimal initialization and makes the parse table shallow where appropriate.

Usage


text_to_flat_pd(text, transformers, more_specs)

Arguments

text
The text to parse.

transformers
Passed to cache_make_key() to generate a key.

more_specs
Passed to cache_make_key() to generate a key.

Details

This includes:

token before and after.

stylerignore attribute.

caching attributes.

Note that the parse table might be shallow if caching is enabled and some values are cached.

[Package styler version 1.10.3 Index]