transform_mixed_non_empty {styler}	R Documentation

Ensure for .Rmd and friends that a code chunk without code is formatted as a code chunk without any lines.

Description

Ensure for .Rmd and friends that a code chunk without code is formatted as a code chunk without any lines.

Usage


transform_mixed_non_empty(r_chunk, transformer_fun)

[Package styler version 1.10.3 Index]