unindent_fun_dec {styler}	R Documentation

Revert the indention of function declaration header

Description

Necessary for consistent indention of the function declaration header.

Usage


unindent_fun_dec(pd, indent_by = 2L)

Arguments

pd
A parse table.

indent_by
How many spaces of indention should be inserted after operators such as '('.

See Also

set_unindention_child update_indention_ref_fun_dec

[Package styler version 1.10.3 Index]