verify_roundtrip {styler}	R Documentation

Verify the styling

Description

If scope was set to "line_breaks" or lower (compare tidyverse_style()), we can compare the expression before and after styling and return an error if it is not the same. If that's not possible, a weaker guarantee that we want to give is that the resulting code is parsable.

Usage


verify_roundtrip(old_text, new_text, parsable_only = FALSE)

Arguments

old_text
The initial expression in its character representation.

new_text
The styled expression in its character representation.

parsable_only
If we should only check for the code to be parsable.

Limitation

Note that this method ignores roxygen code examples and comments and no verification can be conducted if tokens are in the styling scope.

Examples


styler:::verify_roundtrip("a+1", "a + 1")
styler:::verify_roundtrip("a+1", "a + 1 # comments are dropped")
try(styler:::verify_roundtrip("a+1", "b - 3"))

[Package styler version 1.10.3 Index]