visit_one {styler}	R Documentation

Transform a flat parse table with a list of transformers

Description

Uses Reduce() to apply each function of funs sequentially to pd_flat.

Usage


visit_one(pd_flat, funs)

Arguments

pd_flat
A flat parse table.

funs
A list of transformer functions.

See Also

Other visitors: visit

[Package styler version 1.10.3 Index]