wrap_expr_in_expr {styler}	R Documentation

Wrap an expression into an expression

Description

Takes a parse table and wraps it in a new parse table that contains the expression as a child.

Usage


wrap_expr_in_expr(pd)

Arguments

pd
A parse table.

[Package styler version 1.10.3 Index]