wrap_if_else_while_for_fun_multi_line_in_curly {styler}	R Documentation

Wrap if-else, while and for statements in curly braces

Description

Wrap statements in curly braces if it is not already wrapped in a such.

Usage


wrap_if_else_while_for_fun_multi_line_in_curly(pd, indent_by = 2L)

Arguments

pd
A parse table.

indent_by
The amount of spaces used to indent an expression in curly braces. Used for unindention.

[Package styler version 1.10.3 Index]