wrap_subexpr_in_curly {styler}	R Documentation

Wrap a sub-expression in curly braces

Description

Wraps some rows of a parse table into a sub-expression.

Usage


wrap_subexpr_in_curly(pd, ind_to_be_wrapped, indent_by, space_after)

Arguments

pd
A parse table.

ind_to_be_wrapped
The indices of the rows that should be wrapped into a new expression.

space_after
How many spaces should be inserted after the closing brace.

[Package styler version 1.10.3 Index]