Add information about previous / next token to each terminal

Note that this does function must be called in compute_parse_data_nested() and we cannot wait to initialize this attribute until apply_transformers(), where all other attributes are initialized with default_style_guide_attributes() (when using tidyverse_style()) because for cached code, we don't build up the nested structure and leave it shallow (to speed up things), see also shallowify().

Usage

add_terminal_token_after(pd_flat)

add_terminal_token_before(pd_flat)

add_attributes_caching(pd_flat, transformers, more_specs)

Arguments

pd_flat

A flat parse table.

transformers

Passed to cache_make_key() to generate a key.

more_specs

Passed to cache_make_key() to generate a key.

Functions

add_attributes_caching(): Initializes newlines and lag_newlines.