Relocate the expressions containing the token EQ_ASSIGN within the nested parse table

Although syntactically identical, utils::getParseData() does not produce the same hierarchy of the parse table (parent and id relationship) for <- and = (See 'Examples'). This is considered to be a bug and causes problems because the nested parse table constructed with compute_parse_data_nested() is not consistent if EQ_ASSIGN occurs in the expression to style. In particular, EQ_ASSIGN and the tokens to its left and right are located too high up in the hierarchy of the nested parse data. Hence, this function wraps the sub-expression into an expression, similar to wrap_expr_in_curly(). Since wrap_expr_in_curly() is called from within a visitor (and relocate_eq_assign() not), we need to wrap the the implementation relocate_eq_assign_nest() that operates on nests into a visitor call.

relocate_eq_assign(pd)

Arguments

pd
A parse table.

Examples

styler:::get_parse_data("a <- b <- 3")

#> # A tibble: 10 x 11
#>    line1  col1 line2  col2    id parent token    terminal text     pos_id short 
#>    <int> <int> <int> <int> <int>  <int> <chr>    <lgl>    <chr>     <int> <chr> 
#>  1     1     1     1    11    11      0 expr     FALSE    a <- b …      1 "a <-…
#>  2     1     1     1     1     1      3 SYMBOL   TRUE     a             2 "a"   
#>  3     1     1     1     1     3     11 expr     FALSE    a             3 "a"   
#>  4     1     3     1     4     2     11 LEFT_AS… TRUE     <-            4 "<-"  
#>  5     1     6     1    11    10     11 expr     FALSE    b <- 3        5 "b <-…
#>  6     1     6     1     6     4      6 SYMBOL   TRUE     b             6 "b"   
#>  7     1     6     1     6     6     10 expr     FALSE    b             7 "b"   
#>  8     1     8     1     9     5     10 LEFT_AS… TRUE     <-            8 "<-"  
#>  9     1    11     1    11     7      8 NUM_CON… TRUE     3             9 "3"   
#> 10     1    11     1    11     8     10 expr     FALSE    3            10 "3"   

styler:::get_parse_data("a  = b = 3")

#> # A tibble: 10 x 11
#>    line1  col1 line2  col2    id parent token      terminal text    pos_id short
#>    <int> <int> <int> <int> <int>  <int> <chr>      <lgl>    <chr>    <int> <chr>
#>  1     1     1     1    10    14      0 expr_or_a… FALSE    a  = b…      1 "a  …
#>  2     1     1     1     1     1      3 SYMBOL     TRUE     a            2 "a"  
#>  3     1     1     1     1     3     14 expr       FALSE    a            3 "a"  
#>  4     1     4     1     4     2     14 EQ_ASSIGN  TRUE     =            4 "="  
#>  5     1     6     1    10    13     14 expr_or_a… FALSE    b = 3        5 "b =…
#>  6     1     6     1     6     5      7 SYMBOL     TRUE     b            6 "b"  
#>  7     1     6     1     6     7     13 expr       FALSE    b            7 "b"  
#>  8     1     8     1     8     6     13 EQ_ASSIGN  TRUE     =            8 "="  
#>  9     1    10     1    10     9     10 NUM_CONST  TRUE     3            9 "3"  
#> 10     1    10     1    10    10     13 expr       FALSE    3           10 "3"  

styler:::get_parse_data(
  "x = 5
  if(x >= 5)
  y = TRUE else
  y = FALSE",
)

#> # A tibble: 29 x 11
#>    line1  col1 line2  col2    id parent token    terminal text      pos_id short
#>    <int> <int> <int> <int> <int>  <int> <chr>    <lgl>    <chr>      <int> <chr>
#>  1     1     1     1     5     9      0 expr_or… FALSE    "x = 5"        1 "x =…
#>  2     1     1     1     1     1      3 SYMBOL   TRUE     "x"            2 "x"  
#>  3     1     1     1     1     3      9 expr     FALSE    "x"            3 "x"  
#>  4     1     3     1     3     2      9 EQ_ASSI… TRUE     "="            4 "="  
#>  5     1     5     1     5     5      6 NUM_CON… TRUE     "5"            5 "5"  
#>  6     1     5     1     5     6      9 expr     FALSE    "5"            6 "5"  
#>  7     2     3     4    11    42      0 expr     FALSE    "if(x >=…      7 "if(…
#>  8     2     3     2     4    11     42 IF       TRUE     "if"           8 "if" 
#>  9     2     5     2     5    12     42 '('      TRUE     "("            9 "("  
#> 10     2     6     2    11    19     42 expr     FALSE    "x >= 5"      10 "x >…
#> # … with 19 more rows

styler:::get_parse_data(
  "x <- 5
  if(x >= 5)
  y <- TRUE else
  y <- FALSE",
)

#> # A tibble: 29 x 11
#>    line1  col1 line2  col2    id parent token  terminal text        pos_id short
#>    <int> <int> <int> <int> <int>  <int> <chr>  <lgl>    <chr>        <int> <chr>
#>  1     1     1     1     6     7      0 expr   FALSE    "x <- 5"         1 "x <…
#>  2     1     1     1     1     1      3 SYMBOL TRUE     "x"              2 "x"  
#>  3     1     1     1     1     3      7 expr   FALSE    "x"              3 "x"  
#>  4     1     3     1     4     2      7 LEFT_… TRUE     "<-"             4 "<-" 
#>  5     1     6     1     6     4      5 NUM_C… TRUE     "5"              5 "5"  
#>  6     1     6     1     6     5      7 expr   FALSE    "5"              6 "5"  
#>  7     2     3     4    12    39      0 expr   FALSE    "if(x >= 5…      7 "if(…
#>  8     2     3     2     4    10     39 IF     TRUE     "if"             8 "if" 
#>  9     2     5     2     5    11     39 '('    TRUE     "("              9 "("  
#> 10     2     6     2    11    18     39 expr   FALSE    "x >= 5"        10 "x >…
#> # … with 19 more rows