Relocate the expressions containing the token EQ_ASSIGN within the nested parse table

Although syntactically identical, utils::getParseData() does not produce the same hierarchy of the parse table (parent and id relationship) for <- and = (See 'Examples'). This is considered to be a bug and causes problems because the nested parse table constructed with compute_parse_data_nested() is not consistent if EQ_ASSIGN occurs in the expression to style. In particular, EQ_ASSIGN and the tokens to its left and right are located too high up in the hierarchy of the nested parse data. Hence, this function wraps the sub-expression into an expression, similar to wrap_expr_in_curly(). Since wrap_expr_in_curly() is called from within a visitor (and relocate_eq_assign() not), we need to wrap the the implementation relocate_eq_assign_nest() that operates on nests into a visitor call.

relocate_eq_assign(pd)

Arguments

pd
A parse table.

Examples

styler:::get_parse_data("a <- b <- 3")

#> # A tibble: 10 x 11
#>    line1  col1 line2  col2    id parent token       terminal text  pos_id short
#>  * <int> <int> <int> <int> <int>  <int> <chr>       <lgl>    <chr>  <int> <chr>
#>  1     1     1     1    11    11      0 expr        FALSE    a <-…      1 "a <…
#>  2     1     1     1     1     1      3 SYMBOL      TRUE     a          2 a    
#>  3     1     1     1     1     3     11 expr        FALSE    a          3 a    
#>  4     1     3     1     4     2     11 LEFT_ASSIGN TRUE     <-         4 <-   
#>  5     1     6     1    11    10     11 expr        FALSE    b <-…      5 "b <…
#>  6     1     6     1     6     4      6 SYMBOL      TRUE     b          6 b    
#>  7     1     6     1     6     6     10 expr        FALSE    b          7 b    
#>  8     1     8     1     9     5     10 LEFT_ASSIGN TRUE     <-         8 <-   
#>  9     1    11     1    11     7      8 NUM_CONST   TRUE     3          9 3    
#> 10     1    11     1    11     8     10 expr        FALSE    3         10 3    

styler:::get_parse_data("a  = b = 3")

#> # A tibble: 8 x 11
#>   line1  col1 line2  col2    id parent token     terminal text  pos_id short
#> * <int> <int> <int> <int> <int>  <int> <chr>     <lgl>    <chr>  <int> <chr>
#> 1     1     1     1     1     1      3 SYMBOL    TRUE     a          1 a    
#> 2     1     1     1     1     3      0 expr      FALSE    a          2 a    
#> 3     1     4     1     4     2      0 EQ_ASSIGN TRUE     =          3 =    
#> 4     1     6     1     6     4      6 SYMBOL    TRUE     b          4 b    
#> 5     1     6     1     6     6      0 expr      FALSE    b          5 b    
#> 6     1     8     1     8     5      0 EQ_ASSIGN TRUE     =          6 =    
#> 7     1    10     1    10     7      8 NUM_CONST TRUE     3          7 3    
#> 8     1    10     1    10     8      0 expr      FALSE    3          8 3    

styler:::get_parse_data(
  "x = 5
  if(x >= 5)
  y = TRUE else
  y = FALSE",
)

#> # A tibble: 26 x 11
#>    line1  col1 line2  col2    id parent token     terminal text    pos_id short
#>  * <int> <int> <int> <int> <int>  <int> <chr>     <lgl>    <chr>    <int> <chr>
#>  1     1     1     1     1     1      3 SYMBOL    TRUE     x            1 x    
#>  2     1     1     1     1     3      0 expr      FALSE    x            2 x    
#>  3     1     3     1     3     2      0 EQ_ASSIGN TRUE     =            3 =    
#>  4     1     5     1     5     4      5 NUM_CONST TRUE     5            4 5    
#>  5     1     5     1     5     5      0 expr      FALSE    5            5 5    
#>  6     2     3     4    11    41      0 expr      FALSE    "if(x …      6 "if(…
#>  7     2     3     2     4    11     41 IF        TRUE     if           7 if   
#>  8     2     5     2     5    12     41 '('       TRUE     (            8 (    
#>  9     2     6     2    11    19     41 expr      FALSE    x >= 5       9 "x >…
#> 10     2     6     2     6    13     15 SYMBOL    TRUE     x           10 x    
#> # ... with 16 more rows

styler:::get_parse_data(
  "x <- 5
  if(x >= 5)
  y <- TRUE else
  y <- FALSE",
)

#> # A tibble: 29 x 11
#>    line1  col1 line2  col2    id parent token       terminal text  pos_id short
#>  * <int> <int> <int> <int> <int>  <int> <chr>       <lgl>    <chr>  <int> <chr>
#>  1     1     1     1     6     7      0 expr        FALSE    x <-…      1 "x <…
#>  2     1     1     1     1     1      3 SYMBOL      TRUE     x          2 x    
#>  3     1     1     1     1     3      7 expr        FALSE    x          3 x    
#>  4     1     3     1     4     2      7 LEFT_ASSIGN TRUE     <-         4 <-   
#>  5     1     6     1     6     4      5 NUM_CONST   TRUE     5          5 5    
#>  6     1     6     1     6     5      7 expr        FALSE    5          6 5    
#>  7     2     3     4    12    38      0 expr        FALSE    "if(…      7 "if(…
#>  8     2     3     2     4    10     38 IF          TRUE     if         8 if   
#>  9     2     5     2     5    11     38 '('         TRUE     (          9 (    
#> 10     2     6     2    11    18     38 expr        FALSE    x >=…     10 "x >…
#> # ... with 19 more rows