styler-package {styler}	R Documentation

styler: Non-Invasive Pretty Printing of R Code

Description

styler allows you to format .qmd, .R, .Rmd, .Rmarkdown, .Rnw, and/or .Rprofile files, R packages, or entire R source trees according to a style guide. The following functions can be used for styling:

style_text() to style a character vector.

style_file() to style a single file.

style_dir() to style all files in a directory.

style_pkg() to style the source files of an R package.

RStudio Addins to style either selected code or the active file.

Author(s)

Maintainer: Lorenz Walthert lorenz.walthert@icloud.com

Authors:

Kirill Müller kirill@cynkra.com (ORCID)

Other contributors:

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID) (@patilindrajeets) [contributor]

See Also

Useful links:

https://github.com/r-lib/styler

https://styler.r-lib.org

Report bugs at https://github.com/r-lib/styler/issues

Examples


style_text("call( 1)")
style_text("1    + 1", strict = FALSE)
style_text("a%>%b", scope = "spaces")
style_text("a%>%b; a", scope = "line_breaks")
style_text("a%>%b; a", scope = "tokens")

[Package styler version 1.10.3 Index]