Non-invasive pretty printing of R code

styler allows you to format .R files, packages or entire R source trees according to a style guide. The following functions can be used for styling:

style_text() to style a character vector.

style_file() to style a single .R file.

style_dir() to style all .R files in a directory.

style_pkg() to style the source files of an R package.

styler_addins (RStudio Addins) to style either selected code or the active file.

See also

Author

Maintainer: Lorenz Walthert lorenz.walthert@icloud.com

Authors:

Kirill Müller krlmlr+r@mailbox.org

Examples

style_text("call( 1)")
#> call(1)
style_text("1    + 1", strict = FALSE)
#> 1    + 1
style_text("a%>%b", scope = "spaces")
#> a %>% b
style_text("a%>%b; a", scope = "line_breaks")
#> a %>% b; a
style_text("a%>%b; a", scope = "tokens")
#> a %>% b()
#> a