Turn off styling for parts of the code

Using stylerignore markers, you can temporarily turn off styler. See a few illustrative examples below.

Details

Styling is on by default when you run styler.

To mark the start of a sequence where you want to turn styling off, use # styler: off.

To mark the end of this sequence, put # styler: on in your code. After that line, styler will again format your code.

To ignore an inline statement (i.e. just one line), place # styler: off at the end of the line. Note that inline statements cannot contain other comments apart from the marker, i.e. a line like 1 # comment # styler: off won't be ignored.

To use something else as start and stop markers, set the R options styler.ignore_start and styler.ignore_stop using options(). If you want these settings to persist over multiple R sessions, consider setting them in your R profile, e.g. with usethis::edit_rprofile().

Examples

# as long as the order of the markers is correct, the lines are ignored.
style_text(
  "
  1+1
  # styler: off
  1+1
  # styler: on
  1+1
  "
)

#> 
#> 1 + 1
#>   # styler: off
#>   1+1
#> # styler: on
#> 1 + 1

# if there is a stop marker before a start marker, styler won't be able
# to figure out which lines you want to ignore and won't ignore anything,
# issuing a warning.
if (FALSE) {
style_text(
  "
  1+1
  # styler: off
  1+1
  # styler: off
  1+1
  "
)
}