Update indention information of parse data

Update indention information of parse data

indent_round(pd, indent_by)

indent_curly(pd, indent_by)

indent_op(pd, indent_by, token = c(math_token, "SPECIAL-PIPE"))

indent_assign(pd, indent_by, token = c("LEFT_ASSIGN",
  "\n                                                   EQ_ASSIGN"))

indent_without_paren(pd, indent_by = 2)

Arguments

pd
A nested or flat parse table that is already enhanced with line break and space information via create_filler().

indent_by
How many spaces should be added after the token of interest.

token
The token the indention should be based on.

Functions

indent_round: Inserts indetion based on round brackets.

indent_assign: Same as indent_op, but only indents one token after token, not all remaining.

indent_without_paren: Is used to indent if / while / for statements that do not have curly brackets.